+++ src/Main.java
@@ -3,7 +3,8 @@ import java.io.IOException;
 public class Main {
     public static void main(String[] args) {
         try {
-            Shape shape = ShapeBuilder.build();
+            ShapeBuilder shapeBuilder = new ShapeBuilder();
+            Shape shape = shapeBuilder.build();
             System.out.println("Area: " + shape.getArea());
             System.out.println("Perimeter: " + shape.getPerimeter());
         } catch (IOException e) {
