+++ src/ShapeBuilder.java
@@ -4,12 +4,12 @@ import java.io.InputStreamReader;
 
 public class ShapeBuilder {
 
-    public static Shape build() throws IOException {
+    public Shape build() throws IOException {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         System.out.println("Enter figure number:");
         System.out.println("1.Rectangle\n2.Triangle\n3.Circle");
-        Figure figure = Figure.valueOf(Integer.parseInt(reader.readLine()));
-        return switch (figure) {
+        InputShapeType shapeType = InputShapeType.valueOf(Integer.parseInt(reader.readLine()));
+        return switch (shapeType) {
             case RECTANGLE -> createRectangle();
 
             case TRIANGLE -> createTriangle();
@@ -18,7 +18,7 @@ public class ShapeBuilder {
         };
     }
 
-    private static Shape createRectangle() throws IOException {
+    private Shape createRectangle() throws IOException {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         System.out.print("Enter side A: ");
         double sideA = Double.parseDouble(reader.readLine());
@@ -27,7 +27,7 @@ public class ShapeBuilder {
         return new Rectangle(sideA, sideB);
     }
 
-    private static Shape createTriangle() throws IOException {
+    private Shape createTriangle() throws IOException {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         System.out.print("Enter side A: ");
         double sideA = Double.parseDouble(reader.readLine());
@@ -38,7 +38,7 @@ public class ShapeBuilder {
         return new Triangle(sideA, sideB, sideC);
     }
 
-    private static Shape createCircle() throws IOException {
+    private Shape createCircle() throws IOException {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         System.out.print("Enter radius: ");
         double radius = Double.parseDouble(reader.readLine());
